Cover_letter/requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
PyPDF2
